Terminata la parte di trasferimento di tutti i metodi dalla classe DbAndBusiness, necessaria successiva rimozione della classe.
