+++ BusinessLayer/BL_TableManagement.cs
@@ -1,13 +1,10 @@
-﻿using SchoolGrades.DbClasses;
-using System;
+﻿using System;
 using System.Collections.Generic;
-using System.Data.Common;
 using System.Text;
 
 namespace SchoolGrades
 {
     internal partial class BusinessLayer
     {
-        
     }
 }
