+++ BusinessLayer/BusinessLayer.projitems
@@ -17,8 +17,13 @@
     <Compile Include="$(MSBuildThisFileDirectory)BL_ImageManagement.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)BL_LessonManagement.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)BL_LinkManagement.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)BL_PeriodManagement.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)BL_QuestionManagement.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)BL_StudentManagement.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)BL_SubjectManagement.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)BL_TableManagement.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)BL_TagManagement.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)BL_TestManagement.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)BL_UserManagement.cs" />
   </ItemGroup>
 </Project>
