+++ DataLayer/DL_TableManagement.cs
@@ -1,0 +1,78 @@
+﻿using gamon;
+using System;
+using System.Collections.Generic;
+using System.Data;
+using System.Data.Common;
+using System.Data.SQLite;
+using System.Text;
+
+namespace SchoolGrades
+{
+    internal partial class DataLayer
+    {
+        internal void GetLookupTable(string Table, ref DataSet DSet, ref DataAdapter DAdapt)
+        {
+            using (DbConnection conn = Connect())
+            {
+                string query = "SELECT * FROM " + Table + ";";
+                DAdapt = new SQLiteDataAdapter(query, (SQLiteConnection)conn);
+                DSet = new DataSet("OpenLookupTable");
+
+                DAdapt.Fill(DSet);
+                DAdapt.Dispose();
+                DSet.Dispose();
+            }
+        }
+
+        internal void SaveTableOnCvs(DataTable Table, string FileName)
+        {
+            string fileContent = "";
+            foreach (DataColumn col in Table.Columns)
+            {
+                fileContent += col.Caption + '\t';
+            }
+            fileContent += "\r\n";
+            foreach (DataRow row in Table.Rows)
+            {
+                foreach (DataColumn col in Table.Columns)
+                {
+                    fileContent += row[col].ToString() + '\t';
+                }
+                fileContent += "\r\n";
+            }
+            TextFile.StringToFile(FileName, fileContent, false);
+        }
+
+        internal void CreateLookupTableRow(string Table, string IdTable, DataRow Row)
+        {
+            // !!!! TODO !!!! GENERALIZZARE A TABELLE CON NOMI DEI CAMPI ARBITRARI E FAR FUNZIONARE !!!!
+            string query;
+            try
+            {
+                // if key field is Integer, this works
+                int iId = (int)Row[0];
+                query = "INSERT INTO " + Table +
+                    " (" + IdTable + ", name, desc)" +
+                    " VALUES (" + iId + ",'" + Row["name"] + "','" + Row["desc"] + "'" +
+                ");";
+            }
+            catch
+            {
+                // if key field wasn't Integer, this other will work 
+                string sId = (string)Row[0];
+                query = "INSERT INTO " + Table +
+                    " (" + IdTable + ", name, desc)" +
+                    " VALUES ('" + sId + "','" + Row["name"] + "','" + Row["desc"] + "'" +
+                ");";
+            }
+            using (DbConnection conn = Connect())
+            {
+                DbCommand cmd = conn.CreateCommand();
+                cmd.CommandText = query;
+
+                cmd.ExecuteNonQuery();
+                cmd.Dispose();
+            }
+        }
+    }
+}
