+++ DataLayer/DataLayer.projitems
@@ -10,6 +10,7 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="$(MSBuildThisFileDirectory)DL_LessonManagement.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)DL_PeriodManagement.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)DL_QuestionManagement.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)DL_AnnotationManagement.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)DL_AnswerManagement.cs" />
@@ -19,6 +20,10 @@
     <Compile Include="$(MSBuildThisFileDirectory)DL_ImageManagement.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)DL_LinkManagement.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)DL_StudentManagement.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)DL_SubjectManagement.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)DL_TableManagement.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)DL_TagManagement.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)DL_TestManagement.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)DL_TopicManagement.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)DL_UserManagement.cs" />
   </ItemGroup>
