+++ DbClasses/DbAndBusiness.cs
@@ -36,54 +36,6 @@ namespace SchoolGrades.DbClasses
             dbName = PathAndFile;
         }
         #endregion
-        
-        internal void FixQuestionInGrade(int? IdGrade)
-        {
-            using (DbConnection conn = dl.Connect())
-            {
-                DbCommand cmd = conn.CreateCommand();
-
-                cmd.CommandText = "UPDATE Grades" +
-                           " Set" +
-                           " isFixed=TRUE" +
-                           " WHERE idGrade=" + IdGrade +
-                           ";";
-                cmd.ExecuteNonQuery();
-                cmd.Dispose();
-            }
-        }
-        internal void RemoveQuestionFromTest(int? IdQuestion, int? IdTest)
-        {
-            using (DbConnection conn = dl.Connect())
-            {
-                DbCommand cmd = conn.CreateCommand();
-                cmd.CommandText = "DELETE FROM Tests_Questions " +
-                    "WHERE IdQuestion=" + IdQuestion +
-                    " AND IdTest='" + IdTest + "'" +
-                    ";";
-                cmd.ExecuteNonQuery();
-                cmd.Dispose();
-            }
-        }
-
-        internal void SaveTableOnCvs(DataTable Table, string FileName)
-        {
-            string fileContent = "";
-            foreach (DataColumn col in Table.Columns)
-            {
-                fileContent += col.Caption + '\t';
-            }
-            fileContent += "\r\n";
-            foreach (DataRow row in Table.Rows)
-            {
-                foreach (DataColumn col in Table.Columns)
-                {
-                    fileContent += row[col].ToString() + '\t';
-                }
-                fileContent += "\r\n";
-            }
-            TextFile.StringToFile(FileName, fileContent, false);
-        }      
 
         #region funzioni generali per i database
         public int nFieldDbDataReader(string NomeCampo, DbDataReader dr)
@@ -113,890 +65,9 @@ namespace SchoolGrades.DbClasses
         #endregion
 
         #region Metodi specifici per questo programma
-        internal int? UpdateAnnotationGroup(StudentAnnotation currentAnnotation, Student currentStudent)
-        {
-            throw new NotImplementedException();
-        }
-        internal void EraseStudentsPhoto(int? IdStudent, string SchoolYear)
-        {
-            using (DbConnection conn =dl.Connect())
-            {
-                DbCommand cmd = conn.CreateCommand();
-                cmd.CommandText = "DELETE FROM StudentsPhotos_Students" +
-                    " WHERE idStudent=" + IdStudent +
-                    " AND idSchoolYear='" + SchoolYear + "'" +
-                    ";";
-                cmd.ExecuteNonQuery();
-                cmd.Dispose();
-            }
-        }    
-
-        private Question GetQuestionFromRow(DbDataReader Row)
-        {
-            Question q = new Question();
-            q.Difficulty = SafeDb.SafeInt(Row["Difficulty"]);
-            q.IdImage = SafeDb.SafeInt(Row["IdImage"]);
-            q.Duration = SafeDb.SafeInt(Row["Duration"]);
-            q.IdQuestion = SafeDb.SafeInt(Row["IdQuestion"]);
-            q.IdQuestionType = SafeDb.SafeString(Row["IdQuestionType"]);
-            q.IdSchoolSubject = SafeDb.SafeString(Row["IdSchoolSubject"]);
-            //q.IdSubject = SafeDb.SafeInt(Row["IdSubject"]);
-            q.IdTopic = SafeDb.SafeInt(Row["IdTopic"]);
-            q.Image = SafeDb.SafeString(Row["Image"]);
-            q.Text = SafeDb.SafeString(Row["Text"]);
-            q.Weight = SafeDb.SafeDouble(Row["Weight"]);
-            q.NRows = SafeDb.SafeInt(Row["nRows"]);
-            q.IsParamount = SafeDb.SafeInt(Row["isParamount"]);
-            ////////q.IsFixed = SafeDb.SafeBool(Row["isFixed"]);
-
-            return q;
-        }
-
-        internal string GetFilePhoto(int? IdStudent, string SchoolYear)
-        {
-            using (DbConnection conn = dl.Connect())
-            {
-                DbCommand cmd = conn.CreateCommand();
-                string query = "SELECT StudentsPhotos.photoPath" +
-                    " FROM StudentsPhotos_Students, StudentsPhotos" +
-                    " WHERE StudentsPhotos_Students.idStudentsPhoto = StudentsPhotos.idStudentsPhoto";
-                if (SchoolYear != null && SchoolYear != "")
-                    query += " AND StudentsPhotos_Students.idSchoolYear='" + SchoolYear + "'";
-                query += " AND StudentsPhotos_Students.idStudent = " + IdStudent + "; ";
-                string NamePath = null;
-                try
-                {
-                    cmd.CommandText = query;
-                    NamePath = (string)cmd.ExecuteScalar();
-                }
-                catch (Exception ex)
-                {
-
-                }
-                cmd.Dispose();
-                return NamePath;
-            }
-        }
-
-        internal bool IsTopicAlreadyTaught(Topic Topic)
-        {
-            using (DbConnection conn = dl.Connect())
-            {
-                DbCommand cmd = conn.CreateCommand();
-                cmd.CommandText = "SELECT idLesson" +
-                    " FROM Lessons_Topics" +
-                    " WHERE idTopic=" + Topic.Id +
-                    " AND idTopic<>0" +
-                    " LIMIT 1; ";
-                var result = cmd.ExecuteScalar();
-                return (result != null);
-            }
-        }
-
-        internal void SaveSubjects(List<SchoolSubject> SubjectList)
-        {
-            foreach (SchoolSubject s in SubjectList)
-            {
-                SaveSubject(s);
-            }
-        }
-
-        internal string SaveSubject(SchoolSubject Subject)
-        {
-            using (DbConnection conn = dl.Connect())
-            {
-                DbCommand cmd = conn.CreateCommand();
-                if (Subject.OldId != "" && Subject.OldId != null)
-                {
-                    cmd.CommandText = "UPDATE SchoolSubjects " +
-                        "SET" +
-                        " Name='" + SqlVal.SqlString(Subject.Name) + "'" +
-                        ",Desc='" + SqlVal.SqlString(Subject.Desc) + "'" +
-                        ",Color='" + SqlVal.SqlInt(Subject.Color) + "'" +
-                        " WHERE idSchoolSubject='" + SqlVal.SqlString(Subject.IdSchoolSubject) + "'" +
-                        ";";
-                }
-                else
-                {
-                    cmd.CommandText = "INSERT INTO SchoolSubjects " +
-                        "(idSchoolSubject, name, desc, color) " +
-                        "Values ('" + SqlVal.SqlString(Subject.IdSchoolSubject) + "','" + SqlVal.SqlString(Subject.Name)
-                        + "','" + SqlVal.SqlString(Subject.Desc) + "'," + SqlVal.SqlInt(Subject.Color) + "" +
-                        ");";
-                }
-                cmd.ExecuteNonQuery();
-                cmd.Dispose();
-            }
-            return Subject.IdSchoolSubject;
-        }
-
-        internal void GetGradeAndStudent(Grade Grade, Student Student)
-        {
-            using (DbConnection conn = dl.Connect())
-            {
-                DbDataReader dRead;
-                DbCommand cmd = conn.CreateCommand();
-                cmd = conn.CreateCommand();
-                cmd.CommandText = "SELECT Grades.*,Students.* FROM Grades" +
-                    " JOIN Students ON Grades.idStudent = Students.idStudent" +
-                    " WHERE Grades.idGrade=" + Grade.IdGrade.ToString() +
-                    ";";
-                dRead = cmd.ExecuteReader();
-                while (dRead.Read())
-                {
-                    Grade = dl.GetGradeFromRow(dRead);
-                    Student = dl.GetStudentFromRow(dRead);
-                    break; // just the first! 
-                }
-                //dRead.Dispose();
-                //cmd.Dispose();
-            }
-        }
-
-        /// <summary>
-        /// Gets all the grades of a students of a specified IdGradeType that are the sons 
-        /// of another grade which has value greater than zero
-        /// </summary>
-        /// <param Name="IdStudent"></param> student's Id
-        internal DataTable GetMicroGradesOfStudentWithMacroOpen(int? IdStudent, string IdSchoolYear,
-            string IdGradeType, string IdSchoolSubject)
-        {
-            using (DbConnection conn = dl.Connect())
-            {
-                // find the macro grade type of the micro grade
-                // TODO take it from a Grade passed as parameter 
-                DbCommand cmd = conn.CreateCommand();
-                cmd.CommandText = "SELECT idGradeTypeParent " +
-                    "FROM GradeTypes " +
-                    "WHERE idGradeType='" + IdGradeType + "'; ";
-                string idGradeTypeParent = (string)cmd.ExecuteScalar();
-
-                string query = "SELECT datetime(Grades.timestamp),Questions.text,Grades.value" +
-                    ",Grades.weight,Grades.IdGrade,Grades.idGradeParent,Grades.cncFactor" +
-                    ",Questions.IdQuestion,Questions.IdQuestionType,Grades.IdSchoolSubject" +
-                    ",Questions.IdTopic,Questions.Image,Questions.Duration,Questions.Difficulty" +
-                    ",Questions.*" +
-                    ",Grades.*" +
-                    " FROM Grades" +
-                    " JOIN Grades AS Parents" +
-                    " ON Grades.idGradeParent=Parents.idGrade" +
-                    " LEFT JOIN Questions" +
-                    " ON Grades.idQuestion=Questions.idQuestion" +
-                    " WHERE Grades.idStudent =" + IdStudent +
-                    " AND Grades.idSchoolYear='" + IdSchoolYear + "'" +
-                    " AND Grades.idGradeType = '" + IdGradeType + "'" +
-                    " AND Grades.idSchoolSubject = '" + IdSchoolSubject + "'" +
-                    " AND Parents.idGradeType = '" + idGradeTypeParent + "'" +
-                    " AND Grades.idGradeParent = Parents.idGrade" +
-                    " AND (Parents.value = 0 OR Parents.value is NULL)" + 
-                    " ORDER BY Grades.timestamp;";
-
-                DataAdapter DAdapt = new SQLiteDataAdapter(query, (SQLiteConnection)conn);
-                DataSet DSet = new DataSet("OpenMicroGrades");
-                DAdapt.Fill(DSet);
-                DAdapt.Dispose();
-                DSet.Dispose();
-                return DSet.Tables[0];
-            }
-        }
-
+        
         
 
-        internal double GetDefaultWeightOfGradeType(string IdGradeType)
-        {
-            double d;
-            using (DbConnection conn = dl.Connect())
-            {
-                DbCommand cmd = conn.CreateCommand();
-                cmd.CommandText = "SELECT defaultWeight " +
-                    "FROM GradeTypes " +
-                    "WHERE idGradeType='" + IdGradeType + "'; ";
-                d = (double)cmd.ExecuteScalar();
-                cmd.Dispose();
-            }
-            return d;
-        }
-
-        internal SchoolSubject GetSchoolSubject(string IdSchoolSubject)
-        {
-            SchoolSubject subject = new SchoolSubject();
-            DbDataReader dRead;
-            DbCommand cmd;
-            string query;
-            using (DbConnection conn = dl.Connect())
-            {
-                query = "SELECT * FROM SchoolSubjects" +
-                    " WHERE IdSchoolSubject='" + IdSchoolSubject + "'";
-                cmd = new SQLiteCommand(query);
-                cmd.Connection = conn;
-                dRead = cmd.ExecuteReader();
-                while (dRead.Read()) // just the first row 
-                {
-                    subject.Name = dRead["Name"].ToString();
-                    subject.Desc = dRead["Desc"].ToString();
-                    subject.Color = (int)dRead["Color"];
-                    subject.IdSchoolSubject = IdSchoolSubject;
-                    subject.OldId = IdSchoolSubject;
-                }
-            }
-            cmd.Dispose();
-            dRead.Dispose();
-            return subject;
-        }
-
-        internal List<SchoolSubject> GetListSchoolSubjects(bool IncludeANullObject)
-        {
-            List<SchoolSubject> lss = new List<SchoolSubject>();
-            if (IncludeANullObject)
-            {
-                SchoolSubject ss = new SchoolSubject();
-                ss.IdSchoolSubject = "";
-                lss.Add(ss);
-            }
-
-            DbDataReader dRead;
-            DbCommand cmd;
-            string query;
-
-            using (DbConnection conn = dl.Connect())
-            {
-                query = "SELECT * FROM SchoolSubjects";
-                cmd = new SQLiteCommand(query);
-                cmd.Connection = conn;
-                dRead = cmd.ExecuteReader();
-                while (dRead.Read())
-                {
-                    SchoolSubject subject = new SchoolSubject();
-                    subject.IdSchoolSubject = dRead["IdSchoolSubject"].ToString();
-                    subject.OldId = dRead["IdSchoolSubject"].ToString();
-                    subject.Name = dRead["Name"].ToString();
-                    subject.Desc = dRead["Desc"].ToString();
-                    subject.Color = SafeDb.SafeInt(dRead["color"]);
-
-                    lss.Add(subject);
-                }
-            }
-            cmd.Dispose();
-            dRead.Dispose();
-            return lss;
-        }      
-
-        internal void GetLookupTable(string Table, ref DataSet DSet, ref DataAdapter DAdapt)
-        {
-            using (DbConnection conn = dl.Connect())
-            {
-                string query = "SELECT * FROM " + Table + ";";
-                DAdapt = new SQLiteDataAdapter(query, (SQLiteConnection)conn);
-                DSet = new DataSet("OpenLookupTable");
-
-                DAdapt.Fill(DSet);
-                DAdapt.Dispose();
-                DSet.Dispose();
-            }
-        }
-
-        internal void CreateLookupTableRow(string Table, string IdTable, DataRow Row)
-        {
-            // !!!! TODO !!!! GENERALIZZARE A TABELLE CON NOMI DEI CAMPI ARBITRARI E FAR FUNZIONARE !!!!
-            string query;
-            try
-            {
-                // if key field is Integer, this works
-                int iId = (int)Row[0];
-                query = "INSERT INTO " + Table +
-                    " (" + IdTable + ", name, desc)" +
-                    " VALUES (" + iId + ",'" + Row["name"] + "','" + Row["desc"] + "'" +
-                ");";
-            }
-            catch
-            {
-                // if key field wasn't Integer, this other will work 
-                string sId = (string)Row[0];
-                query = "INSERT INTO " + Table +
-                    " (" + IdTable + ", name, desc)" +
-                    " VALUES ('" + sId + "','" + Row["name"] + "','" + Row["desc"] + "'" +
-                ");";
-            }
-            using (DbConnection conn = dl.Connect())
-            {
-                DbCommand cmd = conn.CreateCommand();
-                cmd.CommandText = query;
-
-                cmd.ExecuteNonQuery();
-                cmd.Dispose();
-            }
-        }
-
-        internal List<Tag> GetTagsContaining(string Pattern)
-        {
-            DbDataReader dRead;
-            DbCommand cmd;
-            List<Tag> TagList = new List<Tag>();
-
-            using (DbConnection conn = dl.Connect())
-            {
-                string query = "SELECT *" +
-                    " FROM Tags" +
-                    " WHERE Tag LIKE '%" + SqlVal.SqlString(Pattern) + "%'" +
-                    ";";
-                cmd = new SQLiteCommand(query);
-                cmd.Connection = conn;
-                dRead = cmd.ExecuteReader();
-                while (dRead.Read())
-                {
-                    Tag t = new Tag();
-                    t.IdTag = (int)dRead["IdTag"];
-                    t.TagName = (string)dRead["tag"];
-                    t.Desc = (string)dRead["Desc"];
-
-                    TagList.Add(t);
-                }
-                dRead.Dispose();
-                cmd.Dispose();
-            }
-            return TagList;
-        }
-
-        internal int? CreateNewTag(Tag CurrentTag)
-        {
-            // trova una chiave da assegnare alla nuova domanda
-            CurrentTag.IdTag = dl.NextKey("Tags", "IdTag");
-            using (DbConnection conn = dl.Connect())
-            {
-                DbCommand cmd = conn.CreateCommand();
-                cmd.CommandText = "INSERT INTO Tags " +
-                    "(IdTag, tag, Desc) " +
-                    "Values (" + CurrentTag.IdTag + "," +
-                    "'" + CurrentTag.TagName + "'," +
-                    "'" + CurrentTag.Desc + "'" +
-                    ");";
-                cmd.ExecuteNonQuery();
-                cmd.Dispose();
-            }
-            return CurrentTag.IdTag;
-        }
-
-        internal void SaveTag(Tag CurrentTag)
-        {
-            using (DbConnection conn = dl.Connect())
-            {
-                DbCommand cmd = conn.CreateCommand();
-                cmd.CommandText = "UPDATE Tags " +
-                    " SET IdTag=" + CurrentTag.IdTag + "," +
-                    " tag=" + "'" + CurrentTag.TagName + "'," +
-                    " Desc=" + "'" + CurrentTag.Desc + "'" +
-                    " WHERE idTag=" + CurrentTag.IdTag +
-                    ";";
-                cmd.ExecuteNonQuery();
-                cmd.Dispose();
-            }
-        }
-
-        internal List<Tag> TagsOfAQuestion(int? IdQuestion)
-        {
-            DbDataReader dRead;
-            DbCommand cmd;
-            List<Tag> l = new List<Tag>();
-            using (DbConnection conn = dl.Connect())
-            {
-                string query = "SELECT * " +
-                    " FROM Questions_Tags, Tags" +
-                    " WHERE Tags.IdTag = Questions_Tags.IdTag " +
-                    " AND Questions_Tags.idQuestion=" + IdQuestion +
-                    " ORDER BY Tags.tag;";
-                cmd = new SQLiteCommand(query);
-                cmd.Connection = conn;
-                dRead = cmd.ExecuteReader();
-                while (dRead.Read())
-                {
-                    Tag t = new Tag();
-                    t.Desc = (string)dRead["Desc"];
-                    t.IdTag = (int)dRead["IdTag"];
-                    t.TagName = (string)dRead["tag"];
-                    l.Add(t);
-                }
-                dRead.Dispose();
-                cmd.Dispose();
-            }
-
-            return l;
-        }
-
-        internal void AddTagToQuestion(int? IdQuestion, int? IdTag)
-        {
-            using (DbConnection conn = dl.Connect())
-            {
-                DbCommand cmd = conn.CreateCommand();
-                cmd.CommandText = "INSERT INTO Questions_Tags " +
-                    "(idQuestion, IdTag) " +
-                    "Values (" + IdQuestion + "," +
-                    IdTag +
-                    ");";
-                cmd.ExecuteNonQuery();
-                cmd.Dispose();
-            }
-        }
-
-        internal void DeleteOneStudentFromClass(int? IdDeletingStudent, int? IdClass)
-        {
-            using (DbConnection conn = dl.Connect())
-            {
-                DbCommand cmd = conn.CreateCommand();
-                cmd.CommandText = "DELETE FROM Classes_Students" +
-                    " WHERE Classes_Students.idClass=" + IdClass +
-                    " AND Classes_Students.idStudent=" + IdDeletingStudent.ToString() +
-                    ";";
-                cmd.ExecuteNonQuery();
-                cmd.Dispose();
-            }
-        }
-
-        internal void EraseAllStudentsOfAClass(Class Class)
-        {
-            using (DbConnection conn = dl.Connect())
-            {
-                // erase all the info in tables linked to student
-
-                // erase all the grades of the students of the class 
-                DbCommand cmd = conn.CreateCommand();
-                cmd.CommandText = "DELETE FROM Grades WHERE idStudent IN" +
-                    "(SELECT Students.idStudent FROM Students" +
-                    " JOIN Classes_Students ON Students.idStudent = Classes_Students.idStudent" +
-                    " WHERE Classes_Students.idClass=" + Class.IdClass + ");";
-                cmd.ExecuteNonQuery();
-
-                // erase all the questions of the students of the class
-                cmd.CommandText = "DELETE FROM StudentsQuestions WHERE idStudent IN" +
-                    "(SELECT Students.idStudent FROM Students" +
-                    " JOIN Classes_Students ON Students.idStudent = Classes_Students.idStudent" +
-                    " WHERE Classes_Students.idClass=" + Class.IdClass + ");";
-                cmd.ExecuteNonQuery();
-
-                // erase all the answers of students of the class
-                cmd.CommandText = "DELETE FROM StudentsAnswers WHERE idStudent IN" +
-                    "(SELECT Students.idStudent FROM Students" +
-                    " JOIN Classes_Students ON Students.idStudent = Classes_Students.idStudent" +
-                    " WHERE Classes_Students.idClass=" + Class.IdClass + ");";
-                cmd.ExecuteNonQuery();
-
-                // erase all the tests of students of the class
-                cmd.CommandText = "DELETE FROM StudentsTests WHERE idStudent IN" +
-                    "(SELECT Students.idStudent FROM Students" +
-                    " JOIN Classes_Students ON Students.idStudent = Classes_Students.idStudent" +
-                    " WHERE Classes_Students.idClass=" + Class.IdClass + ");";
-                cmd.ExecuteNonQuery();
-
-                // delete all the photos of students of the class 
-                cmd.CommandText = "DELETE FROM StudentsPhotos WHERE StudentsPhotos.idStudentsPhoto IN" +
-                    "(SELECT StudentsPhotos_Students.idStudentsPhoto" +
-                    " FROM StudentsPhotos, StudentsPhotos_Students, Classes_Students" +
-                    " WHERE StudentsPhotos_Students.idStudent = Classes_Students.idStudent" +
-                    " AND StudentsPhotos.idStudentsPhoto = StudentsPhotos_Students.idStudentsPhoto" +
-                    " AND Classes_Students.idClass=" + Class.IdClass + ");";
-                cmd.ExecuteNonQuery();
-
-                // delete all the references in link table to photos of students of the class 
-                cmd.CommandText = "DELETE FROM StudentsPhotos_Students WHERE idStudent IN" +
-                    "(SELECT StudentsPhotos_Students.idStudent" +
-                    " FROM StudentsPhotos_Students, Classes_Students" +
-                    " WHERE StudentsPhotos_Students.idStudent = Classes_Students.idStudent" +
-                    " AND Classes_Students.idClass=" + Class.IdClass + ");";
-                cmd.ExecuteNonQuery();
-
-                // delete all the students in class
-                // AFTER THIS idStudent OF DELETED IN NOT AVAILABLE ANY LONGER 
-                cmd.CommandText = "DELETE FROM Students WHERE idStudent IN" +
-                    "(SELECT Students.idStudent FROM Students" +
-                    " JOIN Classes_Students ON Students.idStudent = Classes_Students.idStudent" +
-                    " WHERE Classes_Students.idClass=" + Class.IdClass + ");";
-                cmd.ExecuteNonQuery();
-
-                cmd.Dispose();
-            }
-        }
-
-        internal void EraseClassFromClasses(Class Class)
-        {
-            //EraseAllStudentsOfAClass(Class); 
-            using (DbConnection conn = dl.Connect())
-            {
-                // delete all the references in link table between students and classes
-                DbCommand cmd = conn.CreateCommand();
-                cmd.CommandText = "DELETE FROM Classes_Students" +
-                    " WHERE Classes_Students.idClass=" + Class.IdClass +
-                    ";";
-                cmd.ExecuteNonQuery();
-                // erase class from Classes_SchoolSubjects
-                cmd.CommandText = "DELETE FROM Classes_SchoolSubjects" +
-                    " WHERE Classes_SchoolSubjects.idClass=" + Class.IdClass +
-                    ";";
-                cmd.ExecuteNonQuery();
-                // erase class from Classes_Tests
-                cmd.CommandText = "DELETE FROM Classes_Tests" +
-                    " WHERE Classes_Tests.idClass=" + Class.IdClass +
-                    ";";
-                cmd.ExecuteNonQuery();
-                // erase class from table Classes 
-                cmd.CommandText = "DELETE FROM Classes" +
-                    " WHERE Classes.idClass=" + Class.IdClass +
-                    ";";
-                cmd.ExecuteNonQuery();
-                cmd.Dispose();
-            }
-        }
-
-        internal string CreateOneClassOnlyDatabase(Class Class)
-        {
-            string newDatabasePathName = Class.PathRestrictedApplication + "\\SchoolGrades\\Data\\";
-            if (!Directory.Exists(newDatabasePathName))
-                Directory.CreateDirectory(newDatabasePathName);
-
-            string newDatabaseFullName = newDatabasePathName +
-                System.DateTime.Now.ToString("yyyy-MM-dd_HH.mm.ss") +
-                "_" + Class.Abbreviation + "_" + Class.SchoolYear + "_" +
-                Commons.FileDatabase;
-            File.Copy(Commons.PathAndFileDatabase, newDatabaseFullName);
-
-            // open a local connection to database 
-            DataLayer newDatabaseDl = new DataLayer(newDatabaseFullName); 
-
-            // erase all the data of the students of other classes
-            using (DbConnection conn = newDatabaseDl.Connect())
-            {
-                DbCommand cmd = conn.CreateCommand();
-
-                // erase all the other classes
-                cmd.CommandText = "DELETE FROM Classes" +
-                " WHERE idClass<>" + Class.IdClass + ";";
-                cmd.ExecuteNonQuery();
-
-                // erase all the lessons of other classes
-                cmd = conn.CreateCommand();
-                cmd.CommandText = "DELETE FROM Lessons" +
-                    " WHERE idClass<>" + Class.IdClass + ";";
-                cmd.ExecuteNonQuery();
-
-                // erase all the students of other classes from the link table
-                cmd = conn.CreateCommand();
-                cmd.CommandText = "DELETE FROM Classes_Students" +
-                 " WHERE idClass<>" + Class.IdClass + ";";
-                cmd.ExecuteNonQuery();
-
-                // erase all the students of other classes 
-                cmd = conn.CreateCommand();
-                cmd.CommandText = "DELETE FROM Students" +
-                    " WHERE idStudent NOT IN" +
-                    " (SELECT idStudent FROM Classes_Students);";
-                cmd.ExecuteNonQuery();
-
-                // erase all the StartLinks of other classes
-                cmd = conn.CreateCommand();
-                cmd.CommandText = "DELETE FROM Classes_StartLinks" +
-                    " WHERE idClass<>" + Class.IdClass + ";";
-                cmd.ExecuteNonQuery();
-
-                // erase all the grades of other classes' students
-                cmd = conn.CreateCommand();
-                cmd.CommandText = "DELETE FROM Grades" +
-                    " WHERE idStudent NOT IN" +
-                    " (SELECT idStudent FROM Classes_Students);";
-                cmd.ExecuteNonQuery();
-
-                // erase all the links to photos of other classes' students
-                // !! retains previous year's photos of this classes students !!
-                cmd = conn.CreateCommand();
-                cmd.CommandText = "DELETE FROM StudentsPhotos_Students" +
-                    " WHERE idStudent NOT IN" +
-                    " (SELECT idStudent FROM Classes_Students);";
-                cmd.ExecuteNonQuery();
-
-                // erase all the annotations of other classes
-                cmd.CommandText = "DELETE FROM StudentsAnnotations" +
-                    " WHERE idStudent NOT IN" +
-                    " (SELECT idStudent FROM Classes_Students)" + 
-                    ";";
-                cmd.ExecuteNonQuery();
-
-                // erase all the photos of other classes' students
-                // !! retains previous year's photos of this classes students !!
-                cmd = conn.CreateCommand();
-                cmd.CommandText = "DELETE FROM StudentsPhotos" +
-                    " WHERE idStudentsPhoto NOT IN" +
-                    " (SELECT idStudentsPhoto FROM StudentsPhotos_Students);";
-                cmd.ExecuteNonQuery();
-
-                // erase all the questions of the students of the other classes
-                // !! StudentsQuestions currently not used !!
-                cmd.CommandText = "DELETE FROM StudentsQuestions" +
-                    " WHERE idStudent NOT IN" +
-                    " (SELECT idStudent FROM Classes_Students);";
-                cmd.ExecuteNonQuery();
-
-                // erase all the answers  of the students of the other classes
-                // !! StudentsAnswers currently not used !!
-                cmd = conn.CreateCommand();
-                cmd.CommandText = "DELETE FROM StudentsAnswers" +
-                " WHERE idStudent NOT IN" +
-                " (SELECT idStudent FROM Classes_Students);";
-                cmd.ExecuteNonQuery();
-
-                // erase all the tests of students of the other classes
-                // !! StudentsTests currently not used !!
-                cmd = conn.CreateCommand();
-                cmd.CommandText = "DELETE FROM StudentsTests" +
-                " WHERE idStudent NOT IN" +
-                " (SELECT idStudent FROM Classes_Students);";
-                cmd.ExecuteNonQuery();
-
-                // erase all the images of other classes' lessons
-                cmd = conn.CreateCommand();
-                cmd.CommandText = "DELETE FROM Lessons_Images" +
-                    " WHERE idLesson NOT IN" +
-                    " (SELECT idLesson from Lessons);";
-                cmd.ExecuteNonQuery();
-
-                // erase all the topics of other classes' lessons
-                cmd = conn.CreateCommand();
-                cmd.CommandText = "DELETE FROM Lessons_Topics" +
-                    " WHERE idLesson NOT IN" +
-                    " (SELECT idLesson from Lessons);";
-                cmd.ExecuteNonQuery();
-
-                // erase all the users
-                cmd = conn.CreateCommand();
-                cmd.CommandText = "DELETE FROM Users" +
-                    ";";
-                cmd.ExecuteNonQuery();
-
-                // copy all the students' photo files that aren't already there or that have a newer date 
-                string query = "SELECT StudentsPhotos.photoPath" +
-                " FROM StudentsPhotos" +
-                " JOIN StudentsPhotos_Students ON StudentsPhotos_Students.idStudentsPhoto = StudentsPhotos.idStudentsPhoto" +
-                " JOIN Classes_Students ON StudentsPhotos_Students.idStudent = Classes_Students.idStudent" +
-                " WHERE Classes_Students.idClass = " + Class.IdClass + "; ";
-                cmd = new SQLiteCommand(query);
-                cmd.Connection = conn;
-                DbDataReader dReader = cmd.ExecuteReader();
-                while (dReader.Read())
-                {
-                    string destinationFile = Class.PathRestrictedApplication + "\\SchoolGrades\\Images\\" + (string)dReader["photoPath"];
-                    if (!Directory.Exists(Path.GetDirectoryName(destinationFile)))
-                    {
-                        Directory.CreateDirectory(Path.GetDirectoryName(destinationFile));
-                    }
-                    if (!File.Exists(destinationFile) ||
-                        File.GetLastWriteTime(destinationFile)
-                        < File.GetLastWriteTime(Commons.PathImages + "\\" + (string)dReader["photoPath"]))
-                        try
-                        {
-                            // destination file not existing or older
-                            File.Copy(Commons.PathImages + "\\" + (string)dReader["photoPath"],
-                                destinationFile);
-                        }
-                        catch { }
-                }
-                // copy all the picture's files that aren't already there or that have a newer date 
-                query = "SELECT Images.imagePath, Classes.pathRestrictedApplication" +
-                    " FROM Images" +
-                    " JOIN Lessons_Images ON Lessons_Images.idImage=Images.idImage" +
-                    " JOIN Lessons ON Lessons_Images.idLesson=Lessons.idLesson" +
-                    " JOIN Classes ON Classes.idClass=Lessons.idClass" +
-                    " WHERE Lessons.idClass=" + Class.IdClass +
-                    ";";
-                cmd = new SQLiteCommand(query);
-                cmd.Connection = conn;
-                dReader = cmd.ExecuteReader();
-                while (dReader.Read())
-                {
-                    if (dReader["pathRestrictedApplication"] is DBNull)
-                    {
-                        Console.Beep();
-                        break; 
-                    }
-                    if (dReader["imagePath"] is DBNull)
-                    {
-                        Console.Beep();
-                        break;
-                    }
-                    string destinationFile = (string)dReader["pathRestrictedApplication"] +
-                        "\\SchoolGrades\\" + "Images" + "\\" + (string)dReader["imagePath"];
-                    if (!Directory.Exists(Path.GetDirectoryName(destinationFile)))
-                    {
-                        Directory.CreateDirectory(Path.GetDirectoryName(destinationFile));
-                    }
-                    if (!File.Exists(destinationFile) ||
-                        File.GetLastWriteTime(destinationFile)
-                        < File.GetLastWriteTime(Commons.PathImages + "\\" + (string)dReader["imagePath"]))
-                        // destination file not existing or older
-                        try
-                        {
-                            File.Copy(Commons.PathImages + "\\" + (string)dReader["imagePath"],
-                                destinationFile);
-                        }
-                        catch { }
-                }
-                dReader.Dispose();
-                // compact the database 
-                cmd.CommandText = "VACUUM;";
-                cmd.ExecuteNonQuery();
-
-                cmd.Dispose();
-            }
-            return Class.PathRestrictedApplication;
-        }
-
-      
-        internal List<SchoolPeriod> GetSchoolPeriodsOfDate(DateTime Date)
-        {
-            List<SchoolPeriod> l = new List<SchoolPeriod>();
-            using (DbConnection conn = dl.Connect())
-            {
-                DbDataReader dRead;
-                DbCommand cmd = conn.CreateCommand();
-                cmd.CommandText = "SELECT *" +
-                    " FROM SchoolPeriods" +
-                    " WHERE " + SqlVal.SqlDate(Date) +
-                    " BETWEEN dateStart and dateFinish" +
-                    ";";
-                dRead = cmd.ExecuteReader();
-
-                while (dRead.Read())
-                {
-                    SchoolPeriod p = GetOneSchoolPeriodFromRow(dRead);
-                    l.Add(p);
-                }
-            }
-            return l;
-        }
-
-        internal List<SchoolPeriod> GetSchoolPeriods(string IdSchoolYear)
-        {
-            List<SchoolPeriod> l = new List<SchoolPeriod>();
-            using (DbConnection conn = dl.Connect())
-            {
-                DbDataReader dRead;
-                DbCommand cmd = conn.CreateCommand();
-                cmd.CommandText = "SELECT * " +
-                    "FROM SchoolPeriods " +
-                    "WHERE idSchoolYear=" + IdSchoolYear +
-                    " OR IdSchoolYear IS null OR IdSchoolYear=''" +
-                    ";";
-                dRead = cmd.ExecuteReader();
-
-                while (dRead.Read())
-                {
-                    SchoolPeriod p = GetOneSchoolPeriodFromRow(dRead);
-                    l.Add(p);
-                }
-            }
-            return l;
-        }
-
-        internal SchoolPeriod GetOneSchoolPeriodFromRow(DbDataReader Row)
-        {
-            SchoolPeriod p = new SchoolPeriod();
-            p.IdSchoolPeriodType = SafeDb.SafeString(Row["idSchoolPeriodType"]);
-            if (p.IdSchoolPeriodType != "N")
-            {
-                p.DateFinish = SafeDb.SafeDateTime(Row["dateFinish"]);
-                p.DateStart = SafeDb.SafeDateTime(Row["dateStart"]);
-            }
-            p.Name = SafeDb.SafeString(Row["name"]);
-            p.Desc = SafeDb.SafeString(Row["desc"]);
-            p.IdSchoolPeriod = SafeDb.SafeString(Row["idSchoolPeriod"]);
-            p.IdSchoolYear = SafeDb.SafeString(Row["idSchoolYear"]);
-            return p;
-        }
-        internal Test GetTestFromRow(DbDataReader Row)
-        {
-            Test t = new Test();
-            t.IdTest = SafeDb.SafeInt(Row["idTest"]);
-            t.Name = SafeDb.SafeString(Row["name"]);
-            t.Desc = SafeDb.SafeString(Row["desc"]);
-            t.IdSchoolSubject = SafeDb.SafeString(Row["IdSchoolSubject"]);
-            t.IdTestType = SafeDb.SafeString(Row["IdTestType"]);
-            t.IdSchoolSubject = SafeDb.SafeString(Row["IdSchoolSubject"]);
-            t.IdTopic = SafeDb.SafeInt(Row["IdTopic"]);
-
-            return t;
-        }
-        internal Test GetTest(int? IdTest)
-        {
-            Test t = new Test();
-            using (DbConnection conn = dl.Connect())
-            {
-                DbDataReader dRead;
-                DbCommand cmd = conn.CreateCommand();
-                cmd.CommandText = "SELECT * " +
-                    "FROM Tests " +
-                    "WHERE IdTest=" + IdTest +
-                    ";";
-                dRead = cmd.ExecuteReader();
-
-                while (dRead.Read())
-                {
-                    t = GetTestFromRow(dRead);
-                }
-            }
-            return t;
-        }
-        internal List<Test> GetTests()
-        {
-            List<Test> list = new List<Test>();
-            using (DbConnection conn = dl.Connect())
-            {
-                DbDataReader dRead;
-                DbCommand cmd = conn.CreateCommand();
-                cmd.CommandText = "SELECT * " +
-                    "FROM Tests " +
-                    //"WHERE idSchoolYear=" + IdSchoolYear +
-                    //" OR IdSchoolYear IS null OR IdSchoolYear=''" +
-                    ";";
-                dRead = cmd.ExecuteReader();
-
-                while (dRead.Read())
-                {
-                    Test t = GetTestFromRow(dRead);
-                    list.Add(t);
-                }
-            }
-            return list;
-        }
-
-        internal void SaveTest(Test TestToSave)
-        {
-            using (DbConnection conn = dl.Connect())
-            {
-                DbCommand cmd = conn.CreateCommand();
-                if (TestToSave.IdTest == 0 || TestToSave.IdTest == null)
-                {   // create new record
-                    int nextId = dl.NextKey("Tests", "idTest");
-
-                    cmd.CommandText = "INSERT INTO Tests " +
-                    "(idTest,name,desc,IdSchoolSubject,IdTestType,IdTopic" +
-                    ")" +
-                    "Values " +
-                    "(" + nextId + ",'" + SqlVal.SqlString(TestToSave.Name) + "','" +
-                    SqlVal.SqlString(TestToSave.Desc) + "','" + SqlVal.SqlString(TestToSave.IdSchoolSubject) + "'," +
-                     SqlVal.SqlInt(TestToSave.IdTestType) + "," +  SqlVal.SqlInt(TestToSave.IdTopic) +
-                    ");";
-                }
-                else
-                {   // update old record
-                    cmd.CommandText = "UPDATE Tests" +
-                    " SET name='" + SqlVal.SqlString(TestToSave.Name) + "'," +
-                    "desc='" + SqlVal.SqlString(TestToSave.Desc) + "'" +
-                    ",IdSchoolSubject=" + SqlVal.SqlString(TestToSave.IdSchoolSubject) +
-                    ",IdTestType=" +  SqlVal.SqlInt(TestToSave.IdTestType) +
-                    ",IdTopic=" +  SqlVal.SqlInt(TestToSave.IdTopic) +
-                    ")" +
-                    " WHERE idTest=" + TestToSave.IdTest +
-                    ";";
-                }
-                cmd.ExecuteNonQuery();
-                cmd.Dispose();
-            }
-        }
         
         #endregion
     }
