+++ SchoolGrades/frmTestGrading.cs
@@ -30,7 +30,7 @@ namespace SchoolGrades
 
         private void frmTestAssessment_Load(object sender, EventArgs e)
         {
-            currentTest = db.GetTest(1); //!!!!!!!!!!!!!!
+            currentTest = dl.GetTest(1); //!!!!!!!!!!!!!!
             currentClass = dl.GetClass(Commons.IdSchool, "1920", "IFTS"); //!!!!!!!!!!!!!!
 
             GradeTest();
